this seems to work now
